+++ template/render/render.go
@@ -7,6 +7,7 @@ package render // import "github.com/wabarc/wayback/template/render"
 import (
 	"bytes"
 	"net/url"
+	"sort"
 	"strings"
 	"text/template"
 
@@ -52,7 +53,7 @@ func ForPublish(r Renderer) *Render {
 // String returns a string from the Render.
 func (r *Render) String() string {
 	if r != nil {
-		return r.buf.String()
+		return strings.TrimSpace(r.buf.String())
 	}
 	return ""
 }
@@ -187,4 +188,51 @@ func writeArtifact(cols []wayback.Collect, rdx reduxer.Reduxer, fn func(art redu
 			fn(bundle.Artifact())
 		}
 	}
+}
+
+func original(v interface{}) (o string) {
+	var sm = make(map[string]int)
+	if vv, ok := v.([]wayback.Collect); ok && len(vv) > 0 {
+		for _, col := range vv {
+			sm[col.Src] += 1
+		}
+	} else if vv, ok := v.(*Collects); ok {
+		for _, cols := range *vv {
+			for _, dst := range cols.Dst {
+				for src := range dst {
+					sm[src] += 1
+				}
+			}
+		}
+	} else {
+		return o
+	}
+
+	if len(sm) == 0 {
+		return o
+	}
+
+	type kv struct {
+		Key   string
+		Value int
+	}
+
+	ss := make([]kv, 0, len(sm))
+	for k, v := range sm {
+		ss = append(ss, kv{k, v})
+	}
+	sort.Slice(ss, func(i, j int) bool {
+		return ss[i].Value > ss[j].Value
+	})
+
+	var sb strings.Builder
+	sb.WriteString("• source\n")
+	for _, kv := range ss {
+		sb.WriteString(`> `)
+		sb.WriteString(kv.Key)
+		sb.WriteString("\n")
+	}
+	sb.WriteString("\n————\n")
+
+	return sb.String()
 }