+++ template/render/twitter_test.go
@@ -9,48 +9,44 @@ import (
 	"testing"
 )
 
-func TestRenderTwitter(t *testing.T) {
+func TestRenderTwitterForReply(t *testing.T) {
+	const tweet = `• Internet Archive
+> https://web.archive.org/123/https://example.com/
+> https://web.archive.org/123/https://example.org/
+
+• archive.today
+> http://archive.today/abcdE
+> http://archive.today/abc
+
+#wayback #存档`
+
+	got := ForReply(&Twitter{Cols: multi, Data: bundleExample}).String()
+	if !strings.Contains(got, tweet) {
+		t.Errorf("Unexpected render template for Twitter, got \n%s\ninstead of \n%s", got, tweet)
+	}
+}
+
+func TestRenderTwitterForPublish(t *testing.T) {
 	const tweet = `‹ Example ›
 
-source:
-• https://example.com/
+• source
+> https://example.com/
 
 ————
 
-Internet Archive:
-• https://web.archive.org/web/20211000000001/https://example.com/
+• Internet Archive
+> https://web.archive.org/web/20211000000001/https://example.com/
 
-archive.today:
-• http://archive.today/abcdE
+• archive.today
+> http://archive.today/abcdE
 
-IPFS:
-• https://ipfs.io/ipfs/QmTbDmpvQ3cPZG6TA5tnar4ZG6q9JMBYVmX2n3wypMQMtr
+• IPFS
+> https://ipfs.io/ipfs/QmTbDmpvQ3cPZG6TA5tnar4ZG6q9JMBYVmX2n3wypMQMtr
 
 #wayback #存档`
 
 	got := ForPublish(&Twitter{Cols: collects, Data: bundleExample}).String()
 	if got != tweet {
 		t.Errorf("Unexpected render template for Twitter got \n%s\ninstead of \n%s", got, tweet)
-	}
-}
-
-func TestRenderTwitterForReply(t *testing.T) {
-	const source = `• https://example.org/`
-	const tweet = `Internet Archive:
-• https://web.archive.org/123/https://example.com/
-• https://web.archive.org/123/https://example.org/
-
-archive.today:
-• http://archive.today/abcdE
-• http://archive.today/abc
-
-#wayback #存档`
-
-	got := ForReply(&Twitter{Cols: multi, Data: bundleExample}).String()
-	if !strings.Contains(got, source) {
-		t.Errorf("Unexpected render template for Twitter, got \n%s\ninstead of \n%s", got, source)
-	}
-	if !strings.Contains(got, tweet) {
-		t.Errorf("Unexpected render template for Twitter, got \n%s\ninstead of \n%s", got, tweet)
 	}
 }