+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add support for reads from stdin and files ([#301](https://github.com/wabarc/wayback/pull/301))
+- Add support for publish to Nostr ([#311](https://github.com/wabarc/wayback/pull/311))
+  - Message content styling
 
 ### Changed
 - Sign images using cosign
@@ -29,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fix semgrep scan workflow ([#312](https://github.com/wabarc/wayback/pull/312))
+- Fix terminal determination
 
 ## [0.18.1] - 2022-10-30